+++ src/main/java/com/koddy/server/auth/presentation/dto/request/ConfirmAuthCodeRequest.java
@@ -1,7 +1,9 @@
 package com.koddy.server.auth.presentation.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
+import lombok.Builder;
 
+@Builder
 public record ConfirmAuthCodeRequest(
         @NotBlank(message = "인증 코드는 필수입니다.")
         String authCode
