+++ src/main/java/com/koddy/server/member/presentation/dto/request/CompleteMentorRequest.java
@@ -7,9 +7,11 @@ import com.koddy.server.member.domain.model.mentor.Schedule;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
+import lombok.Builder;
 
 import java.util.List;
 
+@Builder
 public record CompleteMentorRequest(
         @NotBlank(message = "이름은 필수입니다.")
         String name,
