+++ src/main/java/com/koddy/server/member/presentation/dto/request/MentorScheduleRequest.java
@@ -1,9 +1,11 @@
 package com.koddy.server.member.presentation.dto.request;
 
 import com.koddy.server.member.domain.model.mentor.Day;
+import lombok.Builder;
 
 import java.time.LocalTime;
 
+@Builder
 public record MentorScheduleRequest(
         Day day,
         LocalTime startTime,
