+++ src/main/java/com/koddy/server/member/presentation/dto/request/UpdateMentorPasswordRequest.java
@@ -1,7 +1,9 @@
 package com.koddy.server.member.presentation.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
+import lombok.Builder;
 
+@Builder
 public record UpdateMentorPasswordRequest(
         @NotBlank(message = "기존 비밀번호는 필수입니다.")
         String currentPassword,
