+++ src/main/java/com/koddy/server/member/presentation/dto/request/UpdateMentorScheduleRequest.java
@@ -3,9 +3,11 @@ package com.koddy.server.member.presentation.dto.request;
 import com.koddy.server.member.domain.model.mentor.Period;
 import com.koddy.server.member.domain.model.mentor.Schedule;
 import jakarta.validation.constraints.NotEmpty;
+import lombok.Builder;
 
 import java.util.List;
 
+@Builder
 public record UpdateMentorScheduleRequest(
         @NotEmpty(message = "멘토링 스케줄은 하루 이상 선택해야 합니다.")
         List<MentorScheduleRequest> schedules
